refactor(design): move the step contract to src/types/designV2.types.ts
designV2Store imported STEPS and the view types from the components
folder, which inverts the dependency direction (state depending on UI).
The step contract now lives in the shared types layer, next to the other
*.types.ts files, and both the store and the components import it from
there.
